requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for Intercom API
aiohttp==3.9.1

# OpenAI integration
openai==1.3.7

# Vector Database and RAG
chromadb==1.0.15
langchain==0.3.23
langchain-openai==0.3.26
langchain-community==0.3.23
sentence-transformers==2.6.1

# Environment variables and configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# Data validation
pydantic==2.5.0

# CORS middleware
python-multipart==0.0.6

# Logging and monitoring
structlog==23.2.0

# Security and cryptography
cryptography==41.0.8

# Date and time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools (optional)
black==23.11.0
isort==5.12.0
flake8==6.1.0 